+++ app/views/incomes/estimatedPayment/update/checkYourAnswers.scala.html
@@ -51,7 +51,7 @@
 
 @Link.toInternalPage(
     id=Some("estimatedPayLink"),
-    url=income.estimatedPay.update.routes.IncomeUpdateCalculatorController.estimatedPayPage().url,
+    url=income.estimatedPay.update.routes.IncomeUpdateEstimatedPayController.estimatedPayPage().url,
     value=Some(messages("tai.WhatDoYouWantToDo.submit")),
     cssClasses=Some("button")
 ).toHtml
