+++ app/views/incomes/howToUpdate.scala.html
@@ -45,7 +45,7 @@ messages: Messages, user: controllers.auth.AuthedUser, templateRenderer: uk.gov.
 ){
 
     <div class="section flush--bottom">
-        @form(action=income.estimatedPay.update.routes.IncomeUpdateCalculatorController.handleChooseHowToUpdate){
+        @form(action=income.estimatedPay.update.routes.IncomeUpdateHowToUpdateController.handleChooseHowToUpdate){
                 @{inputRadioGroupStyling(
                     field=howToUpdateForm("howToUpdate"),
                     radioOptions=Seq(("incomeCalculator" -> Messages("tai.howToUpdate.incomeCalculator")), ("enterAnnual" -> Messages("tai.howToUpdate.enterAnnual"))),
