+++ app/views/incomes/workingHours.scala.html
@@ -42,7 +42,7 @@
     employerName = Some(employerName)
 ){
     <div class="section flush--bottom soft--top">
-        @form(action=controllers.income.estimatedPay.update.routes.IncomeUpdateCalculatorController.handleWorkingHours){
+        @form(action=controllers.income.estimatedPay.update.routes.IncomeUpdateWorkingHoursController.handleWorkingHours){
                 @{inputRadioGroupStyling(
                     field=workingHoursForm("workingHours"),
                     radioOptions=Seq(("regularHours" -> messages("tai.workingHours.regularHours")), ("irregularHours" -> messages("tai.workingHours.irregularHours"))),
